# Doc/swagger.yaml
openapi: 3.0.3
info:
  title: Green Dii API
  description: |
    REST API for Green Dii Sales Workflow System
    ระบบจัดการ Sales Workflow — Quotation → Work Order → Hand Over → Purchase Request
    **Default password for all seeded users: `1234`**
  version: 1.0.0
  contact:
    name: Green Dii Support
servers:
  - url: http://localhost:4000/api
    description: Development
  - url: https://your-domain.com/api
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:         { type: string }
        username:   { type: string }
        fullName:   { type: string }
        initials:   { type: string }
        role:       { type: string, enum: [sales, sales2, sale_mgr, admin_mgr, project_mgr, director, procurement, factory] }
        lineUserId: { type: string }
        active:     { type: boolean }
        createdAt:  { type: string, format: date-time }
...
    QuotationItem:
      type: object
      properties:
        desc:   { type: string }
        qty:    { type: number }
        unit:   { type: string }
        price:  { type: number }
        amount: { type: number }
    Quotation:
      type: object
      properties:
        id:           { type: string }
        quoNo:        { type: string }
        salesId:      { type: string }
        customerName: { type: string }
        project:      { type: string }
        grandTotal:   { type: number }
        status:       { type: string, enum: [draft, pending, approved, rejected, cancelled] }
        approvalStep: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/QuotationItem' }
        createdAt:    { type: string, format: date-time }
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      responses:
        "200": { description: User list, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } } }
    post:
      tags: [Users]
      summary: Create user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, fullName, role]
              properties:
                username: { type: string }
                password: { type: string }
                fullName: { type: string }
                initials: { type: string }
                role:     { type: string }
      responses:
        "201": { description: User created }
  /users/{id}:
    put:
      tags: [Users]
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:   { type: string }
                lineUserId: { type: string }
                role:       { type: string }
                active:     { type: boolean }
      responses:
        "200": { description: Updated user }
    delete:
      tags: [Users]
      summary: Deactivate user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deactivated }

  # ── CUSTOMERS ────────────────────────────────────────────────────────────────
...
  /units:
    get:
      tags: [Units]
      summary: List units
      responses:
        "200": { description: Units list }
    post:
      tags: [Units]
      summary: Create unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: แผ่น }
      responses:
        "201": { description: Created }
  /units/{id}:
    put:
      tags: [Units]
      summary: Update unit
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:   { type: string }
                active: { type: boolean }
      responses:
        "200": { description: Updated }
    delete:
      tags: [Units]
      summary: Deactivate unit
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Deactivated }

  # ── QUOTATIONS ───────────────────────────────────────────────────────────────
  /quotations:
    get:
      tags: [Quotations]
      summary: List quotations
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [draft, pending, approved, rejected] }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200": { description: Quotation list }
    post:
      tags: [Quotations]
      summary: Create quotation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Quotation' }
      responses:
        "201": { description: Created }
  /quotations/{id}:
    get:
      tags: [Quotations]
      summary: Get quotation detail
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Quotation with items and approval logs }
    put:
      tags: [Quotations]
      summary: Update quotation (draft/rejected only)
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Quotation' }
      responses:
        "200": { description: Updated }
    delete:
      tags: [Quotations]
      summary: Cancel quotation
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Cancelled }
  /quotations/{id}/submit:
    post:
      tags: [Quotations]
      summary: Submit for approval (draft → pending)
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Submitted }
  /quotations/{id}/approve:
    post:
      tags: [Quotations]
      summary: Approve at current step
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment: { type: string }
      responses:
        "200": { description: Approved }
  /quotations/{id}/reject:
    post:
      tags: [Quotations]
      summary: Reject quotation
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment: { type: string }
      responses:
        "200": { description: Rejected }

  # ── WORK ORDERS ──────────────────────────────────────────────────────────────
...
  /upload:
    post:
      tags: [Upload]
      summary: Upload files
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:            { type: array, items: { type: string, format: binary } }
                category:         { type: string }
                quotationId:      { type: string }
                workOrderId:      { type: string }
                handOverJobId:    { type: string }
                purchaseRequestId:{ type: string }
      responses:
        "201": { description: Uploaded file list }
  /upload/{id}:
    delete:
      tags: [Upload]
      summary: Delete file
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Deleted }