# api/src/routes/upload.js
const router = require('express').Router();
const multer = require('multer');
const path = require('path');
const fs = require('fs');
const prisma = require('../lib/prisma');
const { authenticate } = require('../middleware/auth');

const UPLOAD_DIR = path.join(__dirname, '../../uploads');
if (!fs.existsSync(UPLOAD_DIR)) fs.mkdirSync(UPLOAD_DIR, { recursive: true });

const storage = multer.diskStorage({
  destination: (_req, _file, cb) => cb(null, UPLOAD_DIR),
  filename: (_req, file, cb) => {
    const unique = Date.now() + '-' + Math.round(Math.random() * 1e9);
    cb(null, unique + path.extname(file.originalname));
  },
});

const upload = multer({
  storage,
  limits: { fileSize: (Number(process.env.MAX_FILE_SIZE_MB) || 10) * 1024 * 1024 },
  fileFilter: (_req, file, cb) => {
    const allowed = /jpeg|jpg|png|gif|pdf|doc|docx|xls|xlsx|zip|rar/;
    const ext = allowed.test(path.extname(file.originalname).toLowerCase());
    const mime = allowed.test(file.mimetype);
    if (ext || mime) return cb(null, true);
    cb(new Error('File type not allowed'));
  },
});

// POST /api/upload
router.post('/', authenticate, upload.array('files', 10), async (req, res, next) => {
  try {
    const { category, quotationId, workOrderId, handOverJobId, purchaseRequestId } = req.body;
    const saved = [];
    for (const file of req.files || []) {
      const attachment = await prisma.attachment.create({
        data: {
          filename: file.filename,
          originalName: file.originalname,
          mimeType: file.mimetype,
          size: file.size,
          category: category || null,
          quotationId: quotationId || null,
          workOrderId: workOrderId || null,
          handOverJobId: handOverJobId || null,
          purchaseRequestId: purchaseRequestId || null,
        },
      });
      saved.push({ ...attachment, url: `/uploads/${file.filename}` });
    }
    res.status(201).json(saved);
  } catch (e) { next(e); }
});

// DELETE /api/upload/:id
router.delete('/:id', authenticate, async (req, res, next) => {
  try {
    const att = await prisma.attachment.findUniqueOrThrow({ where: { id: req.params.id } });
    const filePath = path.join(UPLOAD_DIR, att.filename);
    if (fs.existsSync(filePath)) fs.unlinkSync(filePath);
    await prisma.attachment.delete({ where: { id: req.params.id } });
    res.json({ message: 'File deleted' });
  } catch (e) { next(e); }
});

module.exports = router;